Fix TestIndexer handleLoad crash on cached-effect cache loads
handleLoad assumed every load targets a registered entity and dereferenced
state.entityConfigs[tableName].table. The cached-effect load passes the
effect-cache table (envio_effect_<name>), which isn't an entity, so the lookup
was undefined and .table threw "Cannot read properties of undefined (reading
'table')". TestIndexer never persists effect caches, so return an empty result
for non-entity tables and let the effect recompute.

Move the reproduction into EventHandler.test.ts (the existing TestIndexer test
file) so all TestIndexer cases live in one place, and compact its comment.
